@@ -1,13 +1,14 @@
-package com.plugin.samusprogressbar;
+package com.plugin.samusprogressbar.integration;
 
 import com.intellij.remoterobot.RemoteRobot;
 import com.intellij.remoterobot.fixtures.ComponentFixture;
 import com.intellij.remoterobot.launcher.Ide;
 import com.intellij.remoterobot.launcher.IdeDownloader;
 import com.intellij.remoterobot.launcher.IdeLauncher;
+import com.intellij.remoterobot.utils.Keyboard;
 import com.plugin.samusprogressbar.pages.DialogFixture;
 import com.plugin.samusprogressbar.pages.IdeaFrame;
-import com.plugin.samusprogressbar.steps.Steps;
+import com.plugin.samusprogressbar.integration.steps.Steps;
 import okhttp3.OkHttpClient;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -19,7 +20,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -28,21 +28,22 @@ import static com.intellij.remoterobot.search.locators.Locators.byXpath;
 import static com.intellij.remoterobot.stepsProcessing.StepWorkerKt.step;
 import static com.intellij.remoterobot.utils.RepeatUtilsKt.waitFor;
 import static com.plugin.samusprogressbar.RemoteRobotExtKt.isAvailable;
-import static com.plugin.samusprogressbar.pages.ActionMenuFixtureKt.actionMenu;
-import static com.plugin.samusprogressbar.pages.ActionMenuFixtureKt.actionMenuItem;
+import static java.awt.event.KeyEvent.VK_ALT;
+import static java.awt.event.KeyEvent.VK_CONTROL;
+import static java.awt.event.KeyEvent.VK_F9;
+import static java.awt.event.KeyEvent.VK_S;
 import static java.time.Duration.ofMinutes;
 import static java.time.Duration.ofSeconds;
 import static org.assertj.swing.timing.Pause.pause;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ProgressBarIntegrationTest {
+class ProgressBarIntegrationTest {
 
+    private static RemoteRobot remoteRobot;
     private static Process ideaProcess;
     private static Path tmpDir;
-    private static RemoteRobot remoteRobot;
-    private final static Ide.BuildType buildType = Ide.BuildType.RELEASE;
-    private final static String version = "2023.3.4";
-    private final Steps sharedSteps = new Steps(remoteRobot);
+    private static final Ide.BuildType buildType = Ide.BuildType.RELEASE;
+    private static final String VERSION = "2025.2";
 
     static {
         try {
@@ -54,9 +55,9 @@ public class ProgressBarIntegrationTest {
 
 
     @BeforeAll
-    public static void before() {
+    static void before() {
         final OkHttpClient client = new OkHttpClient();
-        remoteRobot = new RemoteRobot("http://localhost:8082", client);
+        remoteRobot = new RemoteRobot("http://127.0.0.1:8082");
         final IdeDownloader ideDownloader = new IdeDownloader(client);
         final Map<String, Object> additionalProperties = new HashMap<>();
         additionalProperties.put("robot-server.port", 8082);
@@ -66,9 +67,9 @@ public class ProgressBarIntegrationTest {
         plugins.add(ideDownloader.downloadRobotPlugin(tmpDir));
         plugins.add(Paths.get("src","test","resources", "SamusProgressBar.jar"));
         ideaProcess = IdeLauncher.INSTANCE.launchIde(
-                ideDownloader.downloadAndExtract(Ide.IDEA_COMMUNITY, tmpDir, buildType, version),
+                ideDownloader.downloadAndExtract(Ide.IDEA_COMMUNITY, tmpDir, buildType, VERSION),
                 additionalProperties,
-                Collections.emptyList(),
+                List.of("-Dide.show.tips.on.startup.default.value=false", "-Didea.trust.all.projects=true"),
                 plugins,
                 tmpDir,
                 stringMap -> null
@@ -77,14 +78,16 @@ public class ProgressBarIntegrationTest {
     }
 
     @AfterAll
-    public static void after() throws IOException {
+    static void after() throws IOException {
         ideaProcess.destroy();
         // Disable for the pipeline
         // FileUtils.cleanDirectory(tmpDir.toFile());
     }
 
+
     @Test
-    public void select_a_character_and_launch_build() {
+    void select_a_character_and_launch_build() {
+        Steps sharedSteps = new Steps(remoteRobot);
         sharedSteps.createNewProject();
         sharedSteps.closeTipOfTheDay();
         sharedSteps.waitUntilAllTheBgTasksFinish();
@@ -93,8 +96,8 @@ public class ProgressBarIntegrationTest {
         waitFor(ofMinutes(5), () -> !idea.isDumbMode());
 
         step("Setup the plugin's settings", () -> {
-            actionMenu(remoteRobot, "File").click();
-            actionMenuItem(remoteRobot, "Settings...").click();
+            Keyboard keyboard = new Keyboard(remoteRobot);
+            keyboard.hotKey(VK_CONTROL, VK_ALT, VK_S);
             pause(ofSeconds(2).toMillis());
             final DialogFixture settingsDialog = remoteRobot.find(DialogFixture.class,
                     DialogFixture.byTitle("Settings"), Duration.ofSeconds(10));
@@ -105,8 +108,8 @@ public class ProgressBarIntegrationTest {
         });
 
         step("Trigger the progress bar", () -> {
-            actionMenu(remoteRobot, "Build").click();
-            actionMenuItem(remoteRobot, "Rebuild Project").click();
+            Keyboard keyboard = new Keyboard(remoteRobot);
+            keyboard.hotKey(VK_CONTROL, VK_F9);
             List<ComponentFixture> list = remoteRobot.findAll(ComponentFixture.class, byXpath("//div[@class='JProgressBar']"));
             ComponentFixture componentFixture = list.get(0);
             assertEquals(20, componentFixture.getRemoteComponent().getHeight());
